+++ src/dotnet/APIView/ClientSPA/src/app/_components/index-page/index-page.component.spec.ts
@@ -2,9 +2,6 @@ import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
 
 import { IndexPageComponent } from './index-page.component';
-import { NavBarComponent } from '../shared/nav-bar/nav-bar.component';
-import { ReviewsListComponent } from '../reviews-list/reviews-list.component';
-import { RevisionsListComponent } from '../revisions-list/revisions-list.component';
 import { AppModule } from 'src/app/app.module';
 import { SharedAppModule } from 'src/app/_modules/shared/shared-app.module';
 
@@ -15,10 +12,7 @@ describe('IndexPageComponent', () => {
   beforeEach(() => {
     TestBed.configureTestingModule({
       declarations: [
-        IndexPageComponent,
-        NavBarComponent,
-        ReviewsListComponent,
-        RevisionsListComponent
+        IndexPageComponent
       ],
       imports: [
         HttpClientTestingModule,
@@ -28,7 +22,6 @@ describe('IndexPageComponent', () => {
     });
     fixture = TestBed.createComponent(IndexPageComponent);
     component = fixture.componentInstance;
-    fixture.detectChanges();
   });
 
   it('should create', () => {