+++ src/dotnet/APIView/ClientSPA/src/app/_components/review-page-options/review-page-options.component.ts
@@ -52,7 +52,6 @@ export class ReviewPageOptionsComponent implements OnInit, OnChanges {
   @Output() showHiddenAPIEmitter : EventEmitter<boolean> = new EventEmitter<boolean>();
   @Output() showLeftNavigationEmitter : EventEmitter<boolean> = new EventEmitter<boolean>();
   @Output() disableCodeLinesLazyLoadingEmitter : EventEmitter<boolean> = new EventEmitter<boolean>();
-  @Output() markAsViewedEmitter : EventEmitter<boolean> = new EventEmitter<boolean>();
   @Output() subscribeEmitter : EventEmitter<boolean> = new EventEmitter<boolean>();
   @Output() showLineNumbersEmitter : EventEmitter<boolean> = new EventEmitter<boolean>();
   @Output() apiRevisionApprovalEmitter : EventEmitter<boolean> = new EventEmitter<boolean>();
@@ -74,7 +73,6 @@ export class ReviewPageOptionsComponent implements OnInit, OnChanges {
   showDocumentationSwitch : boolean = true;
   showHiddenAPISwitch : boolean = false;
   showLeftNavigationSwitch : boolean = true;
-  markedAsViewSwitch : boolean = false;
   subscribeSwitch : boolean = false;
   showLineNumbersSwitch : boolean = true;
   disableCodeLinesLazyLoading: boolean = false;
@@ -170,12 +168,10 @@ export class ReviewPageOptionsComponent implements OnInit, OnChanges {
 
     if (changes['userProfile'] && changes['userProfile'].currentValue != undefined) {
       this.setSubscribeSwitch();
-      this.setMarkedAsViewSwitch();
       this.setPageOptionValues();
     }
 
     if (changes['activeAPIRevision'] && changes['activeAPIRevision'].currentValue != undefined) {
-      this.setMarkedAsViewSwitch();
       this.selectedApprovers = this.activeAPIRevision!.assignedReviewers.map(reviewer => reviewer.assingedTo);
       this.isCopilotReviewSupported = this.isCopilotReviewSupportedForPackage();
       this.setAPIRevisionApprovalStates();
@@ -294,15 +290,7 @@ export class ReviewPageOptionsComponent implements OnInit, OnChanges {
   }
 
   /**
-  * Callback for markedAsViewSwitch Change
-  * @param event the Filter event
-  */
-  onMarkedAsViewedSwitchChange(event: InputSwitchChangeEvent) {
-    this.markAsViewedEmitter.emit(event.checked);
-  }
-
-  /**
-  * Callback for markedAsViewSwitch Change
+  * Callback for subscribeSwitch Change
   * @param event the Filter event
   */
   onSubscribeSwitchChange(event: InputSwitchChangeEvent) {
@@ -463,10 +451,6 @@ export class ReviewPageOptionsComponent implements OnInit, OnChanges {
 
   setSubscribeSwitch() {
     this.subscribeSwitch = (this.userProfile && this.review) ? this.review!.subscribers.includes(this.userProfile?.email!) : this.subscribeSwitch;
-  }
-
-  setMarkedAsViewSwitch() {
-    this.markedAsViewSwitch = (this.activeAPIRevision && this.userProfile)? this.activeAPIRevision!.viewedBy.includes(this.userProfile?.userName!): this.markedAsViewSwitch;
   }
 
   copyReviewText(event: Event) {