+++ src/dotnet/APIView/ClientSPA/src/app/_models/userPreferenceModel.ts
@@ -17,7 +17,6 @@ export class UserPreferenceModel {
     showDocumentation: boolean
     showComments: boolean
     showSystemComments: boolean
-    useBetaIndexPage: boolean
     disableCodeLinesLazyLoading: boolean
     scrollBarSize: ScrollBarSize
 
@@ -35,7 +34,6 @@ export class UserPreferenceModel {
         this.showDocumentation = false;
         this.showComments = true;
         this.showSystemComments = true;
-        this.useBetaIndexPage = false;
         this.disableCodeLinesLazyLoading = false;
         this.scrollBarSize = ScrollBarSize.Small;
     }